Remove presentation module
